(maint) replace weird exception testing with thrown+?
slingshot.test provides a thrown+? test helper, which makes the
expression of these tests much more clear, so switch to that.

Note these will cause eastwood to make the following observation:

    suspicious-expression: and called with 1 args.  (and x) always returns x.  Perhaps there are misplaced parentheses?

Because essentially these tests are transformed to:

    (and (= error_type (:type expected)))

We can silence this by also testing for the message related to the
type, though this may mean more churn in the code.