CTH-328 remove add-hops, add {set,get}{-json,}-debug helpers
Here we remove add-hop and the handling of the special :_hops member, and
instead introduce get-debug and set-debug handlers for interacting with the
debug frame.

Addiotionally  we rework some of the internal structures so a Message
maintains :_chunks in a way that's more like the data structures
used by smee/binary which simplifies encode.