(PCP-467) Update tests for globally enabled validations
The use of the `schema.test/validate-schemas` fixture (introduced in the
previous commit) enables schema validations globally during the test
execution, i.e. the validations are enabled even for code for which they
were not enabled previously.
Some tests were not prepared for that and were failing in the new setup.
This commit fixes those tests and also one minor bug discovered while
updating the tests.