(PCP-620) Validate chunk lengths when decoding
Prior to this patch there was no validation of the PC message chunk's
length field which made it possible to attack the PCP broker by sending
a message with a chunk length set to an incorrectly large numbers. As
the broker allocates a buffer for each chunk using taht length, it could
exhaust its available memory upon receipt of such a malicious message.
To prevent that we introduce a check to ensure that the length of any
chunk is not greater than the length of the entire message.