(PCP-600) Inventory updates prerequisites
This commit adds schema definitions to be used in inventory updates.
Additionally it changes the return value of the (uri-wildcard? ) function
to be the exploded Uri in case it is a wildcard Uri or nil otherwise.
This should not cause problems as the truthiness of the newly returned
values is the same is that of originally returned bool values in clojure.